Misuv 4864 cl279 (#76)
* MISUV-4796 CL279 - Add two fields (Annuity/Royalty payments) and display pence values in Tax calculation page

* MISUV-4796 Support decimal values for totalAnnuityPaymentsTaxCharged

+++ app/models/calculation/taxcalculation/TaxCalculation.scala
@@ -24,7 +24,7 @@ case class TaxCalculation(
                            capitalGainsTax: Option[CapitalGainsTax] = None,
                            totalStudentLoansRepaymentAmount: Option[BigDecimal] = None,
                            saUnderpaymentsCodedOut: Option[BigDecimal] = None,
-                           totalAnnuityPaymentsTaxCharged: Option[Int] = None,
+                           totalAnnuityPaymentsTaxCharged: Option[BigDecimal] = None,
                            totalRoyaltyPaymentsTaxCharged: Option[BigDecimal] = None,
                            totalIncomeTaxAndNicsDue: BigDecimal,
                            totalTaxDeducted: Option[BigDecimal] = None,
